requirements.txt
Flask==2.0.1
flask-sqlalchemy==2.5.1
Flask-Migrate==3.1.0
flask-login==0.5.0
flask-wtf==0.15.1
gunicorn==20.1.0
pillow==8.3.1
pytesseract==0.3.8
python-dotenv==0.19.0
redis==3.5.3
celery==5.2.3
python-magic==0.4.24
pdf2image==1.16.0
Werkzeug==2.0.3
click==8.0.3
SQLAlchemy==1.4.44
